Add a Windows installer
NSIS-based.
The installer creates a Start Menu shortcut to start MPC-QT and adds an
entry to uninstall it in Add/Remove Programs.

The 32x32 size icon is needed for the NSIS installer. The other sizes
ensure the icon looks good for any screen scale.
"magick -transparent white" is needed to make the icon background
transparent.
Since the icon format alpha is 1 bit, and for some reason, converting
the SVG results in a smaller border near the transparent background if
the SVG size is bigger.
The checksum and ZIP weren't used and the installer already does a CRC
check.